pyproject.toml
[tool.poetry]
name = "tg-bot-trading"
version = "0.1.0"
description = ""
authors = ["Vyacheslav <nodegopher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.3.0"
tinkoff-investments = "^0.2.0b97"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
